% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_api_access_functions.R
\name{eia_api}
\alias{eia_api}
\title{EIA URL}
\usage{
eia_api()
}
\value{
string specifying the EIA API base url.
}
\description{
Function specifying the EIA API url, used inside other functions in this package
}
\details{
url chunks to be used in API calls
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
