% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_bulk_data_functions.R
\name{eia_bulk_browse}
\alias{eia_bulk_browse}
\title{EIA browse bulk dataset cache}
\usage{
eia_bulk_browse(fields = c("category_id", "data_set", "name",
  "description"), update_cache = FALSE)
}
\arguments{
\item{fields}{Case-insensitive list of EIA bulk dataset field names. Includes \code{category_id},
\code{data_set}, \code{name} and \code{description} by default.}

\item{update_cache}{Logical expression, if FALSE (default), use the cached list of available RBA
tables (\code{rba_cachelist}), if TRUE, update the list of available datasets.}
}
\value{
data frame of selected
}
\description{
Function to browse selected fields of the bulk dataset cachelist
}
\details{
Browse EIA bulk dataset cachelist

Other available field names include: \code{category_id}, \code{identifier},
  \code{title}, \code{keyword}, \code{publisher}, \code{person}, \code{mbox}, \code{accessLevel},
  \code{accessLevelComment}, \code{accessURL}, \code{webService}, \code{format}, \code{temporal},
  \code{modified}, and \code{spatial} (see \code{\link{eia_bulk_cachelist}}).
}
\examples{
  \donttest{
    eia_bulk_browse()
    eia_bulk_browse(fields=c("category_id", "data_set", "name"))
  }
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
