% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_bulk_data_functions.R
\name{eia_bulk_data}
\alias{eia_bulk_data}
\title{Bulk download data from the EIA API}
\usage{
eia_bulk_data(dataset, simplify = TRUE, data_only = TRUE,
  update_cache = FALSE)
}
\arguments{
\item{dataset}{EIA bulk download data set identifier. This argument will accept a regular
expression search string and download all matching bulk download data sets. If
\code{dataset}="All", all bulk download data files will be downloaded and returned in an R data
frame.}

\item{simplify}{Logical. If \code{TRUE}, the function returns EIA data in data frame objects. If
\code{FALSE}, function returns raw JSON formatted source.}

\item{data_only}{Logical. If \code{TRUE} (the default), returns only EIA series data, in a data
frame.  If \code{FALSE} it returns data and category classification information. (Presently
only applies to simplified data.)}

\item{update_cache}{Logical expression, if FALSE (default), use the cached list of available
ABS.Stat datasets, if TRUE, update the list of available datasets.}
}
\value{
Returns either a data frame with EIA bulk series data or a list containing EIA bulk
  series data and associatioed catalogue information.
}
\description{
This function accesses EIA bulk files. (The EIA bulk download facility does not
  require a registered user API key.)
}
\examples{
  \donttest{
    ## Fetch Annual Energy Outlook 2019 bulk data
    eia_ieo <- eia_bulk_data(dataset="AEO.2019");
  }
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
