% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_bulk_data_functions.R
\name{eia_bulk_manifest}
\alias{eia_bulk_manifest}
\title{EIA Bulk Download URL}
\usage{
eia_bulk_manifest(url = eia_bulk_url())
}
\arguments{
\item{path}{EIA bulk download page path}
}
\value{
string specifying the EIA bulk download url.
}
\description{
Function to download the list of bulk EIA files
}
\details{
Get EIA bulk file manifest
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
