% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_bulk_data_functions.R
\name{eia_bulk_search}
\alias{eia_bulk_search}
\title{EIA bulk data search}
\usage{
eia_bulk_search(pattern, fields = c("data_set", "name", "description",
  "keyword"), ignore.case = TRUE, update_cache = FALSE)
}
\arguments{
\item{pattern}{Character string or regular expression to be matched.}

\item{fields}{Character vector of column names through which to search. By default, the function
searches 'data_set' -- dataset identifier, 'name' -- series name, 'description' -- series
description and 'keyword'.}

\item{ignore.case}{Case sensitive pattern match or not.}

\item{update_cache}{Logical expression, if FALSE (default), use the cached list of available RBA
tables (\code{rba_cachelist}), if TRUE, update the list of available datasets.}
}
\value{
data frame of selected
}
\description{
Function to list bulk EIA datasets matching the specified pattern
}
\details{
Browse EIA bulk dataset manifest

Other available field names include: \code{category_id}, \code{identifier},
  \code{title}, \code{keyword}, \code{publisher}, \code{person}, \code{mbox}, \code{accessLevel},
  \code{accessLevelComment}, \code{accessURL}, \code{webService}, \code{format}, \code{temporal},
  \code{modified}, and \code{spatial} (see \code{\link{eia_bulk_cachelist}}).
}
\examples{
  \donttest{
    eia_bulk_search(pattern = "energy outlook 2019")
    eia_bulk_search(pattern = "annual energy outlook")
    eia_bulk_search(pattern = "international energy outlook")
  }
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
