% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_bulk_data_functions.R
\name{eia_bulk_url}
\alias{eia_bulk_url}
\title{EIA Bulk Download URL}
\usage{
eia_bulk_url(path = "opendata/bulkfiles.php")
}
\arguments{
\item{path}{EIA bulk download page path.}
}
\value{
string specifying the EIA bulk download url.
}
\description{
Function specifying the EIA url, used inside other functions in this package
}
\details{
EIA bulk file download page
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
