% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_query_functions.R
\name{eia_category_query}
\alias{eia_category_query}
\title{Gets name and id for a single category, and also lists its children categories' names and
  ids.}
\usage{
eia_category_query(category_id, api_key, format = "json",
  simplify = TRUE)
}
\arguments{
\item{category_id}{A unique numerical ID of the EIA data category to fetch. If NULL, the API's
root category is returned.}

\item{api_key}{Required. A valid API key is required and may be obtained from EIA registration
page (see details).}

\item{format}{API return format. Valid values are 'xml' or 'json'. Default is for API to return
JSON formatted output. Note that it is also necessary to set \code{simplify = FALSE} to access
'xml' or 'json' formatted output.}

\item{simplify}{Logical. If \code{TRUE}, the function returns data in a data frame. If
\code{FALSE}, function returns data in specified output \code{format}.}
}
\description{
This function requests and returns data from the EIA API.
}
\details{
# Download category data from the EIA API

Users of the EIA API are required to obtain an API Key from the EIA Open Data
  registration page (\url{https://www.eia.gov/opendata/register.php}).  A valid email address is
  required as part of the registration process.

  Note that the number of series in a single request is limited to 100.
}
\examples{
  \donttest{
    eia_cat <- eia_category_query(api_key = "Your API Key")
  }
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
