% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_query_functions.R
\name{eia_geoset_query}
\alias{eia_geoset_query}
\title{Gets a set of the series belonging to a specified geoset_id.}
\arguments{
\item{geoset_id}{A case-sensitive string consisting of letters, numbers, dashes ("-") and
periods (".") that uniquely identifies an EIA geoset.}

\item{region_ids}{A vector of region identifiers to be included in the geoset request.}

\item{api_key}{Required. A valid API key is required and may be obtained from EIA registration
page (see details).}

\item{num}{Number of values to be returns, to be used in conjunction with the \code{end_date}
parameter to return the n-values from an end point. The EIA API will only accept the \code{num}
parameter in conjunction with the \code{end_date} parameter and not
\code{start_date}. If \code{num} is supplied, the \code{start_date} argument will be ignored.}

\item{start_date}{Optional start date filter---if supplied, returns data after the specified
start date. Argument accepts valid numeric, character, date or datetime format object. If
numeric it must be in \%Y form (i.e. four digit year).  If character string, it must be in
year-month-date (\%Y-\%m-\%d) format.}

\item{end_date}{Optional end date filter---if supplied, returns data up to and including the
specified end date. Argument accepts valid numeric, character, date or datetime format
object. If numeric it must be in \%Y form (i.e. four digit year).  If character string, it must
be in year-month-date (\%Y-\%m-\%d) format.}

\item{simplify}{Logical. If \code{TRUE}, the function returns data in a data frame. If
\code{FALSE}, function returns data in specified output \code{format}.}

\item{format}{API return format. Valid values are 'xml' or 'json'. Default is for API to return
JSON formatted output. Only applies if \code{simplify = FALSE}.}
}
\description{
-- TO BE COMPLETED --
}
\details{
# Download EIA Geoset series IDs

Users of the EIA API are required to obtain an API Key from the EIA Open Data
  registration page (\url{https://www.eia.gov/opendata/register.php}).  A valid email address is
  required as part of the registration process.

  Note that the number of series in a single request is limited to 100.
}
\examples{
  \donttest{
    eia_cat <- eia_geoset_query(geoset_id="ELEC.GEN.ALL-99.M", region_ids=c("USA"),
                                api_key = "Your API Key")
  }
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
