% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_api_access_functions.R
\name{eia_get}
\alias{eia_get}
\title{Call EIA API and return results}
\usage{
eia_get(api_query)
}
\arguments{
\item{api_query}{EIA API format query string}

\item{...}{other arguments to httr::build_url.}
}
\value{
json contents of page information
}
\description{
Function that calls the EIA API and returns results
}
\details{
Call the EIA API and return result
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
