% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_api_access_functions.R
\name{eia_search}
\alias{eia_search}
\title{EIA API search data query}
\usage{
eia_search(..., page_num = NULL, rows_per_page = NULL,
  simplify = TRUE)
}
\arguments{
\item{...}{any valid EIA API search argument. Valid search arguments are: \code{series_id},
\code{name} and \code{last_updated}. The function will return results for every search string
and/or every search argument where multiple strings or multiple search arguments supplied---see
\link{Examples} for relevant examples.}

\item{page_num}{default page number (default: NULL).}

\item{rows_per_page}{number of rows per page (default: NULL).}
}
\value{
Returns the Series IDs and details as an R data frame.
}
\description{
Function to search the EIA data catalogue
}
\details{
EIA search function

Search terms provided to arguments \code{series_id} and \code{name} must be
  \code{character} objects and \code{last_updated} searches must be date type objects
  (i.e. either \code{Date}, \code{chron} or \code{POSIXt}) specifying a date range to check. Note
  that if only one date value is provided, the query will search for all series updated on that
  date.
}
\examples{
  ## Series ID search
  eia_search(series_id = "PET.MB")

  ## Keyword search
  eia_search(name = "crude oil");

  ## Date search
  eia_search(last_updated = as.Date(c("2015-01-01", "2015-01-02")))
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
