% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_query_functions.R
\name{eia_series_category_query}
\alias{eia_series_category_query}
\title{Get category names for selected EIA series ID.}
\usage{
eia_series_category_query(series_id, api_key, format = "json",
  simplify = TRUE)
}
\arguments{
\item{series_id}{A case-sensitive string consisting of letters, numbers, dashes ("-") and
periods (".") that uniquely identifies an EIA series. unique numerical ID of the EIA data
category to fetch.}

\item{api_key}{Required. A valid API key is required and may be obtained from EIA registration
page (see details).}

\item{format}{API return format. Valid values are 'xml' or 'json'. Default is for API to return
JSON formatted output. Note that it is also necessary to set \code{simplify = FALSE} to access
'xml' or 'json' formatted output.}

\item{simplify}{Logical. If \code{TRUE}, the function returns data in a data frame. If
\code{FALSE}, function returns data in specified output \code{format}.}
}
\description{
This function requests and returns data from the EIA API.
}
\details{
# Download category data from the EIA API

Users of the EIA API are required to obtain an API Key from the EIA Open Data
  registration page (\url{https://www.eia.gov/opendata/register.php}).  A valid email address is
  required as part of the registration process.

  Note that the number of series in a single request is limited to 100.
}
\examples{
  \donttest{
    eia_cat <- eia_series_category_query(series_id="ELEC.GEN.ALL-AK-99.A", api_key = "Your API Key")
  }
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
