% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_query_functions.R
\name{eia_series_query}
\alias{eia_series_query}
\title{Download data by series ID from the EIA API}
\usage{
eia_series_query(series_id, api_key, start_date = NULL,
  end_date = NULL, num = NULL, format = "json", simplify = TRUE)
}
\arguments{
\item{series_id}{Case-insensitive character vector of one or more EIA Series ID codes. The EIA
Series ID (also called source key) is a case-insensitive string consisting of letters, numbers,
dashes ("-") and periods (".") that uniquely identifies an EIA series.}

\item{api_key}{Required. A valid API key is required and may be obtained from EIA registration
page (see details).}

\item{start_date}{Optional start date filter---if supplied, returns data after the specified
start date. Argument accepts valid numeric, character, date or datetime format object. If
numeric it must be in \%Y form (i.e. four digit year).  If character string, it must be in
year-month-date (\%Y-\%m-\%d) format.}

\item{end_date}{Optional end date filter---if supplied, returns data up to and including the
specified end date. Argument accepts valid numeric, character, date or datetime format
object. If numeric it must be in \%Y form (i.e. four digit year).  If character string, it must
be in year-month-date (\%Y-\%m-\%d) format.}

\item{num}{Number of values to be returns, to be used in conjunction with the \code{end_date}
parameter to return the n-values from an end point. The EIA API will only accept the \code{num}
parameter in conjunction with the \code{end_date} parameter and not
\code{start_date}. If \code{num} is supplied, the \code{start_date} argument will be ignored.}

\item{format}{API return format. Valid values are 'xml' or 'json'. Default is for API to return
JSON formatted output. Note that it is also necessary to set \code{simplify = FALSE} to access
'xml' or 'json' formatted output.}

\item{simplify}{Logical. If \code{TRUE}, the function returns data in a data frame. If
\code{FALSE}, function returns data in specified output \code{format}.}
}
\description{
This function requests and returns data from the EIA API.
}
\details{
# Download data from the EIA API

Users of the EIA API are required to obtain an API Key from the EIA Open Data
  registration page (\url{https://www.eia.gov/opendata/register.php}).  A valid email address is
  required as part of the registration process.

  Note that the number of series in a single request is limited to 100.
}
\examples{
  \donttest{
    eia_data <- eia_series_query("STEO.BREPUUS.A"),
                    api_key = "Your API Key", num = 100, end_date = "2017-06-30",
                    format = "json")
  }
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
