% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_api_access_functions.R
\name{eia_ua}
\alias{eia_ua}
\title{EIA user agent}
\usage{
eia_ua()
}
\value{
string specifying the EIA user agent.
}
\description{
Function specifying the EIA API user agent, used inside other functions in this package
}
\details{
url chunks to be used in API calls
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
