% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_api_access_functions.R
\name{eia_url}
\alias{eia_url}
\title{EIA URL}
\usage{
eia_url()
}
\value{
string specifying the EIA base url.
}
\description{
Function specifying the EIA url, used inside other functions in this package
}
\details{
url chunks to be used in API calls
}
\author{
David Mitchell <david.pk.mitchell@gmail.com>
}
\keyword{internal}
