% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiastats.R
\docType{package}
\name{eiastats}
\alias{eiastats}
\alias{eiastats-package}
\title{eiastats: Functions for accessing and downloading data from U.S. Energy Information
Administration (EIA, <https://www.eia.gov./>) API.}
\description{
The eiastats package provides structured access to all of the information available from the
U.S. EIA API, including: access to all annual, quarterly, monthly and daily/weekly data, as well
as access to the EIA Data Catalog API.
}
\details{
To learn more about the eiastats package, start with the vignettes:
\code{browseVignettes(package = "eiastats")}
}
